Moved header related files to header folder
